Create CommonInterface
