+++ src/lib.rs
@@ -1,1 +1,2 @@
-pub mod core;
+pub mod rr_core;
+pub mod settings;
