+++ src/main.rs
@@ -1,6 +1,6 @@
+use rustic_rover::rr_core::RusticRover;
+use rustic_rover::settings::ApplicationDefaultSetting;
 use iced::Application;
-use rustic_rover::core::RusticRover;
-use rustic_rover::core::utils::ApplicationDefaultSetting;
 
 fn main()->iced::Result{
     let defalt_set = ApplicationDefaultSetting::new();
