+++ src/rr_core/packet.rs
@@ -1,25 +1,7 @@
-use crate::core::interface::{Packet, DualShock4};
-use crate::core::utils::ComboBox;
+use crate::rr_core::interface::{Packet, DualShock4, AssignController};
+use crate::rr_core::utils::ComboBox;
 
-#[derive(Debug, PartialEq, Clone, Copy, Eq)]
-pub enum AssignController {
-    JoyLeftX,
-    JoyLeftY,
-    JoyRightX,
-    JoyRightY,
-    DPadUp,
-    DPadDown,
-    DPadLeft,
-    DPadRight,
-    BtnCircle,
-    BtnCross,
-    BtnTriangle,
-    BtnCube,
-    BtnL1,
-    BtnR1,
-    BtnL2,
-    BtnR2,
-}
+
 
 #[derive(Clone)]
 pub struct PlusMinus
@@ -34,53 +16,7 @@ impl PlusMinus {
     }
 }
 
-impl std::fmt::Display for AssignController {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{}",
-            match self {
-                AssignController::JoyLeftX=>"Left_Stick_X",
-                AssignController::JoyLeftY=>"Left_Stick_Y",
-                AssignController::JoyRightX=>"Right_Stick_X",
-                AssignController::JoyRightY=>"Right_Stick_Y",
-                AssignController::DPadUp=>"Up_Key",
-                AssignController::DPadDown=>"Down_Key",
-                AssignController::DPadLeft=>"Left_Key",
-                AssignController::DPadRight=>"Right_Key",
-                AssignController::BtnCircle=>"Circle_Button",
-                AssignController::BtnCross=>"Cross_Button",
-                AssignController::BtnCube=>"Cube_Button",
-                AssignController::BtnTriangle=>"Triangle_Button",
-                AssignController::BtnL1=>"L1_Button",
-                AssignController::BtnL2=>"L2_Button",
-                AssignController::BtnR1=>"R1_Button",
-                AssignController::BtnR2=>"R2_Button",
-            }
-        )
-    }
-}
 
-impl AssignController {
-    pub const ALL:[AssignController;16]=[
-        AssignController::JoyLeftX,
-        AssignController::JoyLeftY,
-        AssignController::JoyRightX,
-        AssignController::JoyRightY,
-        AssignController::BtnCircle,
-        AssignController::BtnCross,
-        AssignController::BtnCube,
-        AssignController::BtnTriangle,
-        AssignController::BtnL1,
-        AssignController::BtnL2,
-        AssignController::BtnR1,
-        AssignController::BtnR2,
-        AssignController::DPadUp,
-        AssignController::DPadDown,
-        AssignController::DPadRight,
-        AssignController::DPadLeft
-    ];
-}
 
 pub struct PacketCreator
 {
@@ -153,6 +89,11 @@ impl PacketCreator {
                 self.packet_ =None
             }
         }
+    }
+
+    pub fn packet_view()
+    {
+
     }
 }
