+++ src/rr_core/thread_connection.rs
